Fix: Wrong error message when it should be "You already added this address to wallets" when importing a wallet again #5330
