+++ AlphaWallet/Common/Types/Error.swift
@@ -1,9 +1,43 @@
-// Copyright SIX DAY LLC. All rights reserved.
+// Copyright © 2022 Stormbird PTE. LTD.
 
 import Foundation
 import APIKit
-import JSONRPCKit
-import web3swift
+import AlphaWalletFoundation
+
+extension KeystoreError {
+    public var errorDescription: String? {
+        switch self {
+        case .failedToDeleteAccount:
+            return R.string.localizable.accountsDeleteErrorFailedToDeleteAccount()
+        case .failedToDecryptKey:
+            return R.string.localizable.accountsDeleteErrorFailedToDecryptKey()
+        case .failedToImport(let error):
+            return error.localizedDescription
+        case .duplicateAccount:
+            return R.string.localizable.accountsDeleteErrorDuplicateAccount()
+        case .failedToSignTransaction:
+            return R.string.localizable.accountsDeleteErrorFailedToSignTransaction()
+        case .failedToCreateWallet:
+            return R.string.localizable.accountsDeleteErrorFailedToCreateWallet()
+        case .failedToImportPrivateKey:
+            return R.string.localizable.accountsDeleteErrorFailedToImportPrivateKey()
+        case .failedToParseJSON:
+            return R.string.localizable.accountsDeleteErrorFailedToParseJSON()
+        case .accountNotFound:
+            return R.string.localizable.accountsDeleteErrorAccountNotFound()
+        case .failedToSignMessage:
+            return R.string.localizable.accountsDeleteErrorFailedToSignMessage()
+        case .failedToExportPrivateKey:
+            return R.string.localizable.accountsDeleteErrorFailedToExportPrivateKey()
+        case .failedToExportSeed:
+            return R.string.localizable.accountsDeleteErrorFailedToExportSeed()
+        case .accountMayNeedImportingAgainOrEnablePasscode:
+            return R.string.localizable.keystoreAccessKeyNeedImportOrPasscode()
+        case .userCancelled:
+            return R.string.localizable.keystoreAccessKeyCancelled()
+        }
+    }
+}
 
 public struct UndefinedError: LocalizedError { }
 public struct UnknownError: LocalizedError { }
@@ -25,6 +59,8 @@ extension Error {
             return error.localizedDescription
         case let error as RequestCanceledDueToWatchWalletError:
             return error.localizedDescription
+        case let error as KeystoreError:
+            return error.errorDescription ?? UnknownError().localizedDescription
         case let error as LocalizedError:
             return error.errorDescription ?? UnknownError().localizedDescription
         case let error as NSError:
