+++ AlphaWallet/Transfer/Collectibles/ViewModels/TransferTokenBatchCardsViaWalletAddressViewControllerViewModel.swift
@@ -137,38 +137,3 @@ extension OpenURLError {
         }
     }
 }
-
-extension KeystoreError {
-    public var errorDescription: String? {
-        switch self {
-        case .failedToDeleteAccount:
-            return R.string.localizable.accountsDeleteErrorFailedToDeleteAccount()
-        case .failedToDecryptKey:
-            return R.string.localizable.accountsDeleteErrorFailedToDecryptKey()
-        case .failedToImport(let error):
-            return error.localizedDescription
-        case .duplicateAccount:
-            return R.string.localizable.accountsDeleteErrorDuplicateAccount()
-        case .failedToSignTransaction:
-            return R.string.localizable.accountsDeleteErrorFailedToSignTransaction()
-        case .failedToCreateWallet:
-            return R.string.localizable.accountsDeleteErrorFailedToCreateWallet()
-        case .failedToImportPrivateKey:
-            return R.string.localizable.accountsDeleteErrorFailedToImportPrivateKey()
-        case .failedToParseJSON:
-            return R.string.localizable.accountsDeleteErrorFailedToParseJSON()
-        case .accountNotFound:
-            return R.string.localizable.accountsDeleteErrorAccountNotFound()
-        case .failedToSignMessage:
-            return R.string.localizable.accountsDeleteErrorFailedToSignMessage()
-        case .failedToExportPrivateKey:
-            return R.string.localizable.accountsDeleteErrorFailedToExportPrivateKey()
-        case .failedToExportSeed:
-            return R.string.localizable.accountsDeleteErrorFailedToExportSeed()
-        case .accountMayNeedImportingAgainOrEnablePasscode:
-            return R.string.localizable.keystoreAccessKeyNeedImportOrPasscode()
-        case .userCancelled:
-            return R.string.localizable.keystoreAccessKeyCancelled()
-        }
-    }
-}
