+++ AlphaWalletTests/Common/Types/ErrorTests.swift
@@ -1,0 +1,13 @@
+// Copyright © 2022 Stormbird PTE. LTD.
+
+import XCTest
+@testable import AlphaWallet
+import AlphaWalletFoundation
+
+class ErrorTests: XCTestCase {
+    func testMakeSureErrorMessageDefinedInExtensionAvailableCorrectlyAcrossFrameworks() {
+        //Must be stored as `Error` for test
+        let e: Error = KeystoreError.duplicateAccount
+        XCTAssertEqual(e.prettyError, "You already added this address to wallets")
+    }
+}
